// src/app/model/menu.model.ts
export class MenuModel {
	MODULE_LINK: string;
    name: string;
    url: string;
	isparenturl: number;	
    icon: string;
	cssname: string;
	reportdata: any;
	reportdataext: any;
	url0: string;
	url1: string;
	url2: string;
	url3: string;
	url4: string;
	url5: string;
	url6: string;
	url7: string;
	url8: string;
	url9: string;
	url10: string;
	url11: string;
	url12: string;
	url13: string;
	reportdataext1: any;
	reportdataext2: any;
	reportdataext3: any;
	reportdataext4: any;
	reportdataext5: any;
	reportdataext6: any;
	reportdataext7: any;
	reportdataext8: any;
	reportdataext9: any;
	reportdataext10: any;
	reportdataext11: any;
	reportdataext12: any;
	reportdataext13: any;
	loading: number;
	cols: number;
	rows: number;
	y:  number;
	x: number;
    badge: {
        variant: 'info',
        text: 'NEW'
    };
    number_order;
	children: any;
	reportdate: any;
}